requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
Django==5.0.6
executing==2.0.1
ffmpeg==1.4
filelock==3.14.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.23.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.18.0
python-decouple==3.8
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12
tokenizers==0.19.1
torch==2.2.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.0
twelvelabs==0.1.24
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
